requirements.txt
#описані бібліотеки, які ми використали в нашій минулій лабі
lask
requests
pytest
pylint